feat: rename grpc.client to grpc.clients for future clients
